% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_based_filtering.R
\name{correlation_based_filtering}
\alias{correlation_based_filtering}
\title{lallalalallala}
\usage{
correlation_based_filtering(
  X_train,
  y_train,
  MI_threshold,
  cor_threshold,
  X_test
)
}
\arguments{
\item{X_train}{features in training data (Original X_train)}

\item{y_train}{class variable for training set}

\item{MI_threshold}{which is a threshold for the mutual information score between the features (X_train) and the response variable (y_train)}

\item{cor_threshold}{which is a threshold for the linear correlation between the features (X_train)}

\item{X_test}{features in test data (Original X_test)}
}
\value{
a list of lists with three outputs which are the reduced_X_train, reduced_X_test and features_list.
}
\description{
lallalalallala
}
