% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI_analysis_plot.R
\name{MI_analysis_plot}
\alias{MI_analysis_plot}
\title{lallalalallala}
\usage{
MI_analysis_plot(
  MI_threshold_LB,
  MI_threshold_UB,
  MI_threshold_step,
  X_train,
  y_train,
  X_test,
  cor_threshold
)
}
\arguments{
\item{MI_threshold_LB}{which is the lower bound for the selected MI values}

\item{MI_threshold_UB}{which is the upper bound for the selected MI values}

\item{MI_threshold_step}{which is the step to produce MI values}

\item{X_train}{features in training data (original X_train)}

\item{y_train}{class variable for training set}

\item{X_test}{features in test data (original X_test)}

\item{cor_threshold}{which is a threshold for the linear correlation between the features (X_train)}
}
\value{
a data frame containing selected MI_values with their corresponding results for the mean_MI and the number of remaining features.

a 3D plot showing the result in data frame
}
\description{
lallalalallala
}
