% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI_cor_plot.R
\name{MI_cor_plot}
\alias{MI_cor_plot}
\title{lallalalallala}
\usage{
MI_cor_plot(
  MI_threshold_LB,
  MI_threshold_UB,
  MI_threshold_step,
  cor_threshold_LB,
  cor_threshold_UB,
  cor_threshold_step,
  X_train,
  y_train,
  X_test
)
}
\arguments{
\item{MI_threshold_LB}{which is the lower bound for the selected MI values}

\item{MI_threshold_UB}{which is the upper bound for the selected MI values}

\item{MI_threshold_step}{which is the step to produce MI values}

\item{cor_threshold_LB}{which is the lower bound for the selected cor values}

\item{cor_threshold_UB}{which is the upper bound for the selected cor values}

\item{cor_threshold_step}{which is the step to produce cor values}

\item{X_train}{features in training data (original X_train)}

\item{y_train}{class variable for training set}

\item{X_test}{features in test data (original X_test)}
}
\value{
a data frame containing selected MI_values and cor_values with their corresponding results for the number of remaining features.

a 3D plot showing the result in data frame
}
\description{
lallalalallala
}
