% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_train_test.R
\name{scaling_train_test}
\alias{scaling_train_test}
\title{lallalalallala}
\usage{
scaling_train_test(X_train, X_test)
}
\arguments{
\item{X_train}{features in training data (in this project X_train is the reduced_X_train)}

\item{X_test}{features in test data (in this project X_test is the reduced_X_test)}
}
\value{
a list which contains the standardized version of X_train and X_test.
}
\description{
lallalalallala
}
