% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_test.R
\docType{data}
\name{y_test}
\alias{y_test}
\title{A y_test data set}
\format{
A column which contains the class or the response variable in the test set
\describe{
\item{class}{the class variable}
}
}
\usage{
y_test
}
\description{
A y_test data set
}
\keyword{datasets}
