% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_train.R
\docType{data}
\name{y_train}
\alias{y_train}
\title{A y_train data set}
\format{
A column which contains the class or the response variable in the training set
\describe{
\item{class}{the class variable}
}
}
\usage{
y_train
}
\description{
A y_train data set
}
\keyword{datasets}
